feat: Add server unit test and fake database helper
This commit adds a unit test for the server at the root path as well as for /api/v1/data. Additionally, a new file `db-helper.js` was added to the utils folder, which exports a fake DB class and a function to get a MongoDB URI for testing. The class contains sample data for users and reviews.